--- components/Category/graphql/products.graphql
@@ -1,43 +1,0 @@
-#import "~/graphql/_pagination.graphql"
-#import "~/graphql/_price.graphql"
-#import "~/graphql/_filters.graphql"
-
-query ProductsQuery($filters: ProductAttributeFilterInput!, $pageSize: Int = 10, $currentPage: Int = 1) {
-    store: storeConfig {
-        id
-        productUrlSuffix: product_url_suffix
-        categoryUrlSuffix: category_url_suffix
-    }
-
-    products(filter: $filters, pageSize: $pageSize, currentPage: $currentPage) {
-        ...pagination
-
-        ...filters
-
-        items @connection(key: "items") {
-            id
-            title: name
-            urlKey: url_key
-            ...price
-
-            image {
-                alt: label
-                src: url
-            }
-
-            ... on ConfigurableProduct {
-                options: configurable_options {
-                    id
-                    items: values {
-                        id: value_index
-                        label
-                        swatch: swatch_data {
-                            value
-                            __typename
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
