--- components/Filters/index.ts
@@ -1,2 +1,0 @@
-export * from './Filters'
-export { Filters as default } from './Filters'
