+++ components/Products/graphql/filterTypes.graphql
