+++ components/Products/index.ts
@@ -1,0 +1,3 @@
+export * from './Products'
+export * from './useProducts'
+export { Products as default } from './Products'
