+++ components/Search/Search.tsx
@@ -1,19 +1,15 @@
 import React, { FunctionComponent, useMemo, useCallback, useState } from 'react'
 import dynamic from 'next/dynamic'
-import { resolveImage } from '~/lib/resolveImage'
 
-import { Root, TopBar, TopBarWrapper, TopBarFilterButton, FiltersIcon, Content, ProductListWrapper, NoResult, FiltersWrapper } from './Search.styled'
+import { Root, TopBar, TopBarWrapper, TopBarFilterButton, FiltersIcon, NoResult } from './Search.styled'
 
-import { useSearch } from './useSearch'
 import { useNetworkStatus } from '~/hooks/useNetworkStatus'
 import { useRouter } from 'next/router'
-import { useFetchMoreOnScrolling } from '@storystore/ui/dist/hooks/useFetchMoreOnScrolling'
 
-import Link from '~/components/Link'
 import Head from '~/components/Head'
 import SearchBar from '@storystore/ui/dist/components/SearchBar'
-import ProductList from '@storystore/ui/dist/components/ProductList'
-import Filters from '~/components/Filters'
+import Products, { useProducts } from '~/components/Products'
+import Icon from '@storystore/ui/dist/components/Icon'
 
 const Error = dynamic(() => import('~/components/Error'))
 
@@ -24,11 +20,11 @@ export const Search: FunctionComponent<SearchProps> = () => {
 
     const { query = '' } = history.query
 
-    const { queries, api } = useSearch({ queryString: query?.toString() })
+    const _products = useProducts({ search: query.toString() })
 
-    const products = queries.search.data?.products
-
-    const productUrlSuffix = queries.search.data?.store?.productUrlSuffix ?? ''
+    const {
+        queries: { products, filters },
+    } = _products
 
     const online = useNetworkStatus()
 
@@ -38,34 +34,17 @@ export const Search: FunctionComponent<SearchProps> = () => {
         setShowFilter(!showFilters)
     }, [showFilters, setShowFilter])
 
-    /**
-     * Infinite Scroll Effect
-     */
-    useFetchMoreOnScrolling({ threshold: 400, loading: queries.search.loading, hasNextPage: products?.pagination && products.pagination.current < products.pagination.total }, () => {
-        queries.search
-            .fetchMore({
-                variables: {
-                    currentPage: products.pagination.current + 1, // next page
-                },
-                updateQuery: (prev: any, { fetchMoreResult }) => {
-                    if (!fetchMoreResult) return prev
-                    return {
-                        ...prev,
-                        products: {
-                            ...prev.products,
-                            ...fetchMoreResult.products,
-                            items: [...prev.products.items, ...fetchMoreResult.products.items],
-                        },
-                    }
-                },
-            })
-            .catch(() => {})
-    })
-
     const handleOnNewSearch = useCallback(
         async (newQuery: string) => {
             if (newQuery.length === 0 || newQuery.length > 2) {
-                await history.push(`/search?query=${newQuery}`, `/search?query=${newQuery}`, { shallow: true })
+                await history.push({
+                    pathname: history.pathname,
+                    query: {
+                        ...history.query,
+                        query: newQuery,
+                    },
+                })
+
                 window.scrollTo(0, 0)
             }
         },
@@ -73,12 +52,13 @@ export const Search: FunctionComponent<SearchProps> = () => {
     )
 
     const productsCount = useMemo(() => {
-        if (!products) return
-        const { count = 0 } = products
+        const data = products.data?.products
+        if (!data) return
+        const { count = 0 } = data
         return `${count > 999 ? '+999' : count} ${count === 0 || count > 1 ? 'results' : 'result'}`
     }, [products])
 
-    if (!online && !products) return <Error type="Offline" fullScreen />
+    if (!online && !products.data?.products) return <Error type="Offline" fullScreen />
 
     return (
         <React.Fragment>
@@ -87,79 +67,20 @@ export const Search: FunctionComponent<SearchProps> = () => {
             <Root>
                 <TopBar>
                     <TopBarWrapper $margin>
-                        <SearchBar loading={queries.search.loading} label="Search" count={productsCount} value={query.toString()} onUpdate={handleOnNewSearch} />
+                        <SearchBar loading={products.loading} label="Search" count={productsCount} value={query.toString()} onUpdate={handleOnNewSearch} />
 
-                        {queries.search.data?.products?.filters && (
-                            <TopBarFilterButton as="button" type="button" onClick={handleToggleFilters}>
-                                <span>
-                                    <FiltersIcon aria-label="Filters" />
-                                </span>
-                            </TopBarFilterButton>
-                        )}
+                        <TopBarFilterButton as="button" type="button" onClick={handleToggleFilters}>
+                            <span>
+                                <Icon svg={FiltersIcon} aria-label="Filters" count={filters.data?.count} />
+                            </span>
+                        </TopBarFilterButton>
                     </TopBarWrapper>
                 </TopBar>
-                <Content $showFilters={showFilters}>
-                    <ProductListWrapper>
-                        <ProductList
-                            loadingMore={queries.search.loading}
-                            items={products?.items
-                                ?.filter((x: any) => x !== null) // patches results returning nulls. I'm looking at you Gift Cards
-                                .map(({ id, image, price, title, urlKey, options }: any, index: number) => {
-                                    return {
-                                        _id: `${id}--${index}`,
-                                        as: Link,
-                                        href: `/${urlKey + productUrlSuffix}`,
-                                        urlResolver: {
-                                            type: 'PRODUCT',
-                                            id,
-                                            urlKey,
-                                        },
-                                        image: {
-                                            alt: image.alt,
-                                            src: {
-                                                desktop: resolveImage(image.src, { width: 1260 }),
-                                                mobile: resolveImage(image.src, { width: 960 }),
-                                            },
-                                            width: 1274,
-                                            height: 1580,
-                                        },
-                                        price: {
-                                            label: price.maximum.regular.value > price.minimum.regular.value ? 'Starting at' : undefined,
-                                            regular: price.minimum.regular.value,
-                                            special: price.minimum.discount.amountOff && price.minimum.final.value - price.minimum.discount.amountOff,
-                                            currency: price.minimum.regular.currency,
-                                        },
-                                        title: {
-                                            text: title,
-                                        },
-                                        colors: options
-                                            ?.find(({ items }: any) => !!items.find(({ swatch }: any) => swatch.__typename === 'ColorSwatchData'))
-                                            ?.items.map(({ label, swatch }: any) => ({ label, value: swatch.value })),
-                                    }
-                                })}
-                        />
-                    </ProductListWrapper>
-                    <FiltersWrapper>
-                        <Filters
-                            loading={queries.search.loading && queries.search.networkStatus !== 3}
-                            items={queries.search.data?.products?.filters?.map(({ title, code, options }: any) => {
-                                return {
-                                    title,
-                                    code,
-                                    options: options.map(({ count, label, value }: any) => ({
-                                        count,
-                                        label,
-                                        value,
-                                    })),
-                                }
-                            })}
-                            onValues={api.setFilter}
-                        />
-                    </FiltersWrapper>
-                </Content>
+
+                <Products showFilters={showFilters} {..._products} />
             </Root>
 
-            {query && products?.count === 0 && (
+            {query && products.data?.products?.count === 0 && (
                 <NoResult $margin>
                     <Error type="404">
                         We couldn&apos;t find any results for &quot;{query}&quot;. <br />
